move edit controls to be at book level
Edit mode on the individual compoents didn't make sense as the edit is in the domain of a book, not an attibute of a book.  This also had an impact on the wayt that a book is being loaded as the original state, as well as the edited state needs to be maintained to easily/performantly cancel the edits made.